ENH: reuse user config code (PR feedback)
